+++ support/chip-testing/test/cert-framework/tc-idm-4.1-support.test.ts
@@ -146,6 +146,9 @@ class Fixture {
             async parseManualPairingCode(): Promise<never> {
                 throw new InternalError("not used in this test");
             },
+            group: (): never => {
+                throw new InternalError("not used by these tests");
+            },
             node: () => node,
         };